pyproject.toml
[tool.poetry]
name = "fast-api-v1"
version = "0.1.0"
description = ""
authors = ["chandanmaurya0"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.10.1"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
boto3 = "^1.35.90"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
